split up metric & metrics
